Bytter ut Redis-cachene med JCache backet av EhCache. Forholdsvis naiv konfigurasjon, men tror det skal være nok
