--- src/main/kotlin/no/nav/klage/oppgave/config/CacheWithRedisConfiguration.kt
@@ -1,73 +1,0 @@
-package no.nav.klage.oppgave.config
-
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.databind.jsontype.BasicPolymorphicTypeValidator
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
-import com.fasterxml.jackson.module.kotlin.KotlinModule
-import no.nav.klage.oppgave.util.getLogger
-import org.springframework.boot.autoconfigure.cache.RedisCacheManagerBuilderCustomizer
-import org.springframework.cache.annotation.EnableCaching
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.Primary
-import org.springframework.data.redis.cache.RedisCacheConfiguration
-import org.springframework.data.redis.cache.RedisCacheManager.RedisCacheManagerBuilder
-import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer
-import org.springframework.data.redis.serializer.RedisSerializationContext
-import org.springframework.data.redis.serializer.StringRedisSerializer
-import java.time.Duration
-
-@EnableCaching
-@Configuration
-class CacheWithRedisConfiguration {
-
-    companion object {
-
-        const val ENHET_CACHE = "enhet"
-        const val TILGANGER_CACHE = "tilganger"
-        const val ROLLER_CACHE = "roller"
-        const val SAKSBEHANDLERE_I_ENHET_CACHE = "saksbehandlereienhet"
-
-        @Suppress("JAVA_CLASS_ON_COMPANION")
-        private val logger = getLogger(javaClass.enclosingClass)
-    }
-
-    @Bean
-    @Primary
-    fun redisCacheConfiguration(): RedisCacheConfiguration {
-        //Tried a bunch of things, finally found this:
-        //https://stackoverflow.com/questions/52265326/how-can-i-easily-cache-kotlin-objects-in-redis-using-json-via-jackson
-        val objectMapper = ObjectMapper()
-            .registerModule(KotlinModule())
-            .registerModule(JavaTimeModule())
-            .activateDefaultTyping(
-                BasicPolymorphicTypeValidator.builder()
-                    .allowIfBaseType(Any::class.java)
-                    .build(), ObjectMapper.DefaultTyping.EVERYTHING
-            )
-
-        return RedisCacheConfiguration
-            .defaultCacheConfig()
-            .serializeKeysWith(
-                RedisSerializationContext.SerializationPair.fromSerializer(StringRedisSerializer())
-            )
-            .serializeValuesWith(
-                RedisSerializationContext.SerializationPair.fromSerializer(
-                    GenericJackson2JsonRedisSerializer(
-                        objectMapper
-                    )
-                )
-            )
-    }
-
-    @Bean
-    fun myRedisCacheManagerBuilderCustomizer(): RedisCacheManagerBuilderCustomizer? {
-        return RedisCacheManagerBuilderCustomizer { builder: RedisCacheManagerBuilder ->
-            builder.withCacheConfiguration(
-                ROLLER_CACHE,
-                RedisCacheConfiguration.defaultCacheConfig()
-                    .entryTtl(Duration.ofHours(1))
-            )
-        }
-    }
-}
